refactor(stats): wyekstrahuj useMarkAsRead + OwnedUnreadItem + AwardCoverageGrid (Refs #102)
Audyt #2 krok 4/4 — StatsSection NIE był god-componentem (spójny dashboard),
lecz pod-wyekstrahowany: logika oznaczania inline, wiersz posiadane-nieprzeczytane
inline oraz siatka pokrycia nagród z liczeniem procentów w JSX.

- useMarkAsRead (hook): stan markingId/markedIds + POST /api/mark-as-read
  (optymistyczny update sekcji filii / pełny refetch przy "Przeczytane").
- stats/OwnedUnreadItem: wiersz pozycji posiadanej-nieprzeczytanej z akcją.
- stats/AwardCoverageGrid: siatka pokrycia nagród (procent zdobytych).

Zachowanie 1:1. Lint czysty, 270 testów zielonych. Audyt #2 zamknięty.
