+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-example",
   "private": true,
-  "version": "1.15.10",
+  "version": "1.15.11",
   "type": "module",
   "engines": {
     "node": ">=18 <21"
